partial revert of r11675 - causes test failure of TemplatedWMSTileSourceTest.testEPSG3857
git-svn-id: http://josm.openstreetmap.de/svn/trunk@11685 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/data/imagery/AbstractWMSTileSource.java
@@ -28,7 +28,7 @@ public abstract class AbstractWMSTileSource extends TMSTileSource {
     private int[] tileXMax;
     private int[] tileYMax;
     private double[] degreesPerTile;
-    private static final double SCALE_DENOMINATOR_ZOOM_LEVEL_1 = 559082264.0287178d;
+    private static final float SCALE_DENOMINATOR_ZOOM_LEVEL_1 = 559082264.0287178f;
 
     /**
      * Constructs a new {@code AbstractWMSTileSource}.
